requirements.txt
ffmpeg-python
python-mpd2
transliterate